+++ SMC_Project_Tetris++/SMC_Project_Tetris++/Game_2P.h
@@ -29,7 +29,7 @@ private:
 	void CreateShape();
 	void CreateShape_2P();
 	bool IsValidMovement(std::array<Position, 16>);
-	bool IsOccupied(int, int);
+	bool IsOccupied(uint16_t, uint16_t);
 	void ProcessEvents(bool&, uint16_t&);
 	void Render();
 	void FileWriter(std::string file);
